[22기_배승식] 동시성 & 결제 연동 미션 제출합니다.

+++ src/main/java/com/ceos22/cgvclone/domain/snack/dto/UserOrderResponseDTO.java
@@ -3,19 +3,20 @@ package com.ceos22.cgvclone.domain.snack.dto;
 import com.ceos22.cgvclone.domain.snack.entity.Item;
 import com.ceos22.cgvclone.domain.snack.entity.OrderItem;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
 public record UserOrderResponseDTO(
         UUID userOrderUuid,
         List<ItemListDTO> orderedItems,
-        long totalPrice,
+        BigDecimal totalPrice,
         Long theaterId
 ) {
     public static UserOrderResponseDTO of(
             UUID userOrderUuid,
             List<ItemListDTO> orderedItems,
-            long totalPrice,
+            BigDecimal totalPrice,
             Long theaterId
     ) {
         return new UserOrderResponseDTO(
